pyproject.toml
[tool.poetry]
name = "city-infrastructure-platform"
version = "1.3.0"
description = ""
authors = []


[tool.poetry.dependencies]
python = "^3.7"
django = "^3.2.13"

django-auditlog = "^1.0.0"
django-colorfield = "^0.6.3"
django-cors-headers = "^3.11.0"
django-enumfields = "^2.1.1"
django-environ = "^0.8.1"
django-extensions = "^3.1.5"
django-filter = "^21.1"
django-gisserver = { git = "https://github.com/Amsterdam/django-gisserver.git", rev = "4e8ece0bc458129fd247014ec0ec4413a0c2afd6"} # No new versions are released after 1.1.3
django-helusers = "^0.7.1"
django-import-export = "^2.7.1"
django-storages = {extras = ["azure"], version = "^1.12.3"}
djangorestframework = "^3.13.1"
djangorestframework-gis = "^1.0.0"
drf-nested-routers = "^0.93.4"
drf-oidc-auth = "^2.0.0"
drf-spectacular = { git = "https://github.com/tfranzel/drf-spectacular.git", rev = "df9e115ca9b9a3ec84d60df942be28ce3cda13fa"}
psycopg2 = "2.8.6"
python-dateutil = "^2.8.2"
pytz = "^2021.3"
requests = "^2.27.1"
rsa = "^4.6"
sentry-sdk = "^1.5.5"
social-auth-app-django = "^5.0.0"
uWSGI = "^2.0.20"
django-mptt = "^0.13.4"
django-light = "^0.1.0"
docutils = "^0.18.1"


[tool.poetry.dev-dependencies]
autoflake = "^1.4"
autopep8 = "^1.6.0"
black = "^22.1.0"
bump2version = "^1.0.1"
factory-boy = "^3.2.1"
flake8 = "^4.0.1"
importmagic = "^0.1.7"
ipython = "^7.16.3"
isort = "^5.10.1"
jedi = "^0.18.1"
pep8-naming = "^0.12.1"
pydocstyle = "^6.1.1"
pyobjc-core = {version = "^8.2", platform = "darwin"}
pyobjc-framework-Cocoa = {version = "^8.2", platform = "darwin"}
pyobjc-framework-FSEvents = {version = "^8.2", platform = "darwin"}
pytest = "^7.0.1"
pytest-cov = "^3.0.0"
pytest-django = "^4.5.2"
pytest-watch = "^4.2.0"
rope = "^0.22.0"
yapf = "^0.32.0"


[tool.black]
extend-exclude = "migrations"
line_length = 120
